renamed core to base
