+++ project1/kbapp/mutualfunds/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class MutualfundsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "mutualfunds"
+    name = "kbapp.mutualfunds"
